backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.30
pydantic==2.7.1
aiosqlite==0.20.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
Pillow>=10.0.0
PyPDF2>=3.0.0
pytesseract>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
